requirements.txt
# ============================================================================
# VILLA CRÉDITOS API - DEPENDENCIAS
# ============================================================================

# Framework web principal
fastapi>=0.104.0

# Servidor ASGI
uvicorn[standard]>=0.24.0

# Base de datos y ORM
sqlalchemy>=2.0.0

# Validación de datos
pydantic[email]>=2.5.0

# Variables de entorno
python-dotenv>=1.0.0

# Middleware y utilidades
python-multipart>=0.0.6
alembic==1.13.1
